works but planner tends to overflow
