+++ src/renderer/store/sagas/control/formulae/interpolate.ts
@@ -101,7 +101,7 @@ export function* interpolate(
     // todo: write to serial
     // delayDuration && (yield delay(delayDuration)); // todo: uncomment this line
     const stringToWrite = gcodeString ?? '';
-    // yield delay(10);x`x
+    // yield delay(30);
     yield call(serial.write, stringToWrite);
 
     finalValue = { gcodeString };
